move fcn to relevant header
