+++ core/include/GammaComboEngine.h
@@ -15,7 +15,6 @@
 
 class BatchScriptWriter;
 class Combiner;
-class ExternalScanWrapper;
 class FileNameBuilder;
 class MethodBergerBoosScan;
 class MethodCoverageScan;
@@ -133,7 +132,6 @@ class GammaComboEngine {
   void makeLatex(Combiner* c);
   void saveWorkspace(Combiner* c, int i);
   void runToys(Combiner* c);
-  ExternalScanWrapper* createExternalScanner(TString filename, TString label);
 
   OptParser* arg = nullptr;
   std::vector<Combiner*> cmb;
