+++ core/src/GammaComboEngine.cpp
@@ -57,13 +57,6 @@
 #include <boost/algorithm/string.hpp>
 #include <boost/lexical_cast.hpp>
 
-namespace {
-  bool isAngle(const TString& name) {
-    return name.Contains("gamma") || name.Contains("delta") || name.Contains("phi") || name.Contains("theta") ||
-           name.Contains("alpha") || name.Contains("beta");
-  }
-}  // namespace
-
 GammaComboEngine::GammaComboEngine(TString name, int argc, char* argv[]) {
   // time the program
   t.Start();
@@ -872,70 +865,6 @@ void GammaComboEngine::makeAddDelCombinations() {
   }
 }
 
-/// Funtionality for external scan
-ExternalScanWrapper* GammaComboEngine::createExternalScanner(const TString filename, const TString label) {
-  auto info = [](std::string msg) { return Utils::msgBase("GammaComboEngine::createExternalScanner() : ", msg); };
-  auto error = [](std::string msg) {
-    return Utils::errBase("GammaComboEngine::createExternalScanner() : ERROR : ", msg);
-  };
-
-  auto extFile = std::unique_ptr<TFile>(TFile::Open(filename));
-  if (!extFile || extFile->IsZombie()) {
-    error(std::format("Cannot open external scan file {:s}", std::string(filename)));
-    return nullptr;
-  }
-
-  using Utils::getUniqueRootName;
-  if (auto hCL2d = dynamic_cast<TH2F*>(extFile->Get("hCL"))) {
-    info("Loading 2D external scan...");
-
-    // Clone to avoid issues when file closes
-    hCL2d = dynamic_cast<TH2F*>(hCL2d->Clone("hCL_external" + getUniqueRootName()));
-    hCL2d->SetDirectory(0);
-
-    auto hCLs2d = dynamic_cast<TH2F*>(extFile->Get("hCLs"));
-    if (hCLs2d) {
-      hCLs2d = dynamic_cast<TH2F*>(hCLs2d->Clone("hCLs_external" + getUniqueRootName()));
-      hCLs2d->SetDirectory(0);
-    }
-
-    auto extScanner = new ExternalScanWrapper(hCL2d, hCLs2d, "external_" + label, label, arg);
-
-    // Infer from histogram axis names for common angular variables
-    TString xName = hCL2d->GetXaxis()->GetName();
-    TString yName = hCL2d->GetYaxis()->GetName();
-    if (isAngle(xName)) extScanner->getScanVar1()->setUnit("Rad");
-    if (isAngle(yName)) extScanner->getScanVar2()->setUnit("Rad");
-
-    return extScanner;
-
-  } else if (auto hCL1d = dynamic_cast<TH1F*>(extFile->Get("hCL"))) {
-    info("Loading 1D external scan...");
-
-    // Clone to avoid issues when file closes
-    hCL1d = dynamic_cast<TH1F*>(hCL1d->Clone("hCL_external" + getUniqueRootName()));
-    hCL1d->SetDirectory(0);
-
-    auto hCLs1d = dynamic_cast<TH1F*>(extFile->Get("hCLs"));
-    if (hCLs1d) {
-      hCLs1d = dynamic_cast<TH1F*>(hCLs1d->Clone("hCLs_external" + getUniqueRootName()));
-      hCLs1d->SetDirectory(0);
-    }
-
-    auto extScanner = new ExternalScanWrapper(hCL1d, hCLs1d, "external_" + label, label, arg);
-
-    // Infer from histogram axis name for common angular variables
-    TString xName = hCL1d->GetXaxis()->GetName();
-    if (isAngle(xName)) extScanner->getScanVar1()->setUnit("Rad");
-
-    return extScanner;
-
-  } else {
-    error("No TH1F/TH2F `hCL` histogram found");
-    return nullptr;
-  }
-}
-
 ///
 /// print parameter structure of the combinations into
 /// .dot file
@@ -1367,7 +1296,7 @@ void GammaComboEngine::make1dProbPlot(MethodProbScan* scanner, int cId) {
       if (files_len != labels_len) { warning("More external files provided than labels"); }
       for (int i = 1; i < files_len; i++) {
         TString label = (i < labels_len) ? arg->externalScanLabels[i] : TString("External " + std::to_string(i));
-        auto extScanner = createExternalScanner(arg->externalScanFiles[i], label);
+        auto extScanner = createExternalScanner(arg->externalScanFiles[i], label, arg);
         if (extScanner) { ((OneMinusClPlot*)plot)->addScanner(extScanner, 0); }
       }
     }
@@ -1616,7 +1545,7 @@ void GammaComboEngine::make2dProbPlot(MethodProbScan* scanner, int cId) {
     if (files_len != labels_len) { warning("More external files provided than labels"); }
     for (int i = 1; i < files_len; ++i) {
       TString label = (i < labels_len) ? arg->externalScanLabels[i] : TString("External " + std::to_string(i));
-      auto extScanner = createExternalScanner(arg->externalScanFiles[i], label);
+      auto extScanner = createExternalScanner(arg->externalScanFiles[i], label, arg);
       if (extScanner) { ((OneMinusClPlot2d*)plot)->addScanner(extScanner, 0); }
     }
   }
